feat: support binary type

+++ include/rpc_core/serialize/type_ptr.hpp
@@ -1,22 +1,19 @@
 #pragma once
 
 #include <cstdint>
-#include <type_traits>
-
-#include "type_raw.hpp"
 
 namespace rpc_core {
 
 template <typename T, typename std::enable_if<std::is_pointer<T>::value, int>::type = 0>
 inline serialize_oarchive& operator>>(const T& t, serialize_oarchive& oa) {
-  auto ptr = (uint64_t)t;
+  auto ptr = (intptr_t)t;
   ptr >> oa;
   return oa;
 }
 
 template <typename T, typename std::enable_if<std::is_pointer<T>::value, int>::type = 0>
 inline serialize_iarchive& operator<<(T& t, serialize_iarchive& ia) {
-  uint64_t ptr;
+  intptr_t ptr;
   ptr << ia;
   t = (T)ptr;
   return ia;
